+++ M2TWEOP-luaPlugin/luaPlugin/m2tweopHelpers.h
@@ -54,7 +54,7 @@ namespace m2tweopHelpers
 	void unloadTextureFromGame(void* texture);
 
 	void toggleUnitsBMapHighlight();
-	void changeBattlemapCameraDistance(signed short maxZoom);
+	void changeBattlemapCameraDistance(int maxZoom);
 
 	void setConversionLvlFromCastle(int castleLvl, int convertToLvl);
 	void setConversionLvlFromCity(int cityLvl, int convertToLvl);
